fix: refactor operations package to use oclif
