+++ operations/.eslintignore
@@ -1,3 +1,3 @@
 dist
 node_modules
-src/synctx/
+coverage
