+++ operations/package.json
@@ -5,13 +5,13 @@
   "author": "Consensys Software Inc.",
   "license": "Apache-2.0",
   "scripts": {
-    "build": "shx rm -rf dist && tsc -b",
+    "build": "shx rm -rf dist && tsc -p tsconfig.build.json",
     "prettier": "prettier -c '**/*.{js,ts}'",
     "prettier:fix": "prettier -w '**/*.{js,ts}'",
     "lint": "eslint . --ext .ts",
-    "test": "jest --bail --detectOpenHandles --forceExit",
-    "lint:fix": "npm run lint:ts:fix && npm run prettier:fix",
-    "clean": "rimraf node_modules dist",
+    "lint:fix": "eslint . --ext .ts --fix",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest --bail --detectOpenHandles --forceExit",
+    "clean": "rimraf node_modules dist coverage",
     "postpack": "shx rm -f oclif.manifest.json",
     "posttest": "pnpm run lint",
     "prepack": "oclif manifest && oclif readme",
@@ -20,23 +20,16 @@
   "dependencies": {
     "@oclif/core": "4.0.23",
     "@oclif/plugin-help": "6.2.12",
-    "@oclif/plugin-plugins": "5.4.9",
-    "axios": "1.7.2",
-    "ethers": "6.12.0",
-    "oclif": "4.14.35",
-    "yargs": "17.7.2"
+    "@oclif/plugin-plugins": "5.4.10",
+    "axios": "1.7.7",
+    "ethers": "6.13.2"
   },
   "devDependencies": {
     "@jest/globals": "29.7.0",
-    "@oclif/prettier-config": "0.2.1",
     "@oclif/test": "4.0.9",
-    "@types/jest": "29.5.12",
-    "@types/yargs": "17.0.32",
-    "eslint-config-oclif": "5.0.0",
-    "eslint-config-oclif-typescript": "3.0.11",
+    "@types/jest": "29.5.13",
     "jest": "29.7.0",
     "jest-mock-extended": "3.0.5",
-    "nodemon": "3.0.1",
     "shx": "0.3.4",
     "ts-jest": "29.2.5"
   },
@@ -57,12 +50,7 @@
       "@oclif/plugin-help",
       "@oclif/plugin-plugins"
     ],
-    "topicSeparator": " ",
-    "topics": {
-      "eth-transfer": {
-        "description": "Say hello to the world and others"
-      }
-    }
+    "topicSeparator": " "
   },
   "files": [
     "/bin",
