+++ operations/src/utils/common/__tests__/fees.test.ts
@@ -1,7 +1,7 @@
 import { describe, afterEach, jest, it, expect, beforeEach } from "@jest/globals";
 import { JsonRpcProvider } from "ethers";
 import { MockProxy, mock, mockClear } from "jest-mock-extended";
-import { get1559Fees } from "../utils.js";
+import { get1559Fees } from "../fees.js";
 
 describe("Utils", () => {
   let providerMock: MockProxy<JsonRpcProvider>;
