+++ operations/src/utils/common/__tests__/validation.test.ts
@@ -1,56 +1,58 @@
 import { describe, it, expect } from "@jest/globals";
 import { ethers } from "ethers";
-import { isValidUrl, sanitizeAddress, sanitizeETHThreshold, sanitizeHexString, sanitizeUrl } from "../cli.js";
+import { isValidProtocolUrl, validateEthereumAddress, validateHexString, validateUrl } from "../validation.js";
 
-describe("CLI", () => {
-  describe("sanitizeAddress", () => {
+describe("Validation utils", () => {
+  describe("validateEthereumAddress", () => {
     it("should throw an error when the input is not a valid Ethereum address", () => {
       const invalidAddress = "0x0a0e";
-      expect(() => sanitizeAddress("Address")(invalidAddress)).toThrow(`Address is not a valid Ethereum address.`);
+      expect(() => validateEthereumAddress("Address", invalidAddress)).toThrow(
+        `Address is not a valid Ethereum address.`,
+      );
     });
 
     it("should return the input when it is a valid Ethereum address", () => {
       const address = ethers.hexlify(ethers.randomBytes(20));
-      expect(sanitizeAddress("Address")(address)).toStrictEqual(address);
+      expect(validateEthereumAddress("Address", address)).toStrictEqual(address);
     });
   });
 
-  describe("isValidUrl", () => {
+  describe("isValidProtocolUrl", () => {
     it("should return false when the url is not valid", () => {
       const invalidUrl = "www.test.com";
-      expect(isValidUrl(invalidUrl, ["http:", "https:"])).toBeFalsy();
+      expect(isValidProtocolUrl(invalidUrl, ["http:", "https:"])).toBeFalsy();
     });
 
     it("should return false when the url protocol is not allowed", () => {
       const invalidUrl = "tcp://test.com";
-      expect(isValidUrl(invalidUrl, ["http:", "https:"])).toBeFalsy();
+      expect(isValidProtocolUrl(invalidUrl, ["http:", "https:"])).toBeFalsy();
     });
 
     it("should return true when the url valid", () => {
       const url = "http://test.com";
-      expect(isValidUrl(url, ["http:", "https:"])).toBeTruthy();
+      expect(isValidProtocolUrl(url, ["http:", "https:"])).toBeTruthy();
     });
   });
 
-  describe("sanitizeUrl", () => {
+  describe("validateUrl", () => {
     it("should throw an error when the input is not a valid url", () => {
       const invalidUrl = "www.test.com";
-      expect(() => sanitizeUrl("Url", ["http:", "https:"])(invalidUrl)).toThrow(
+      expect(() => validateUrl("Url", invalidUrl, ["http:", "https:"])).toThrow(
         `Url, with value: ${invalidUrl} is not a valid URL`,
       );
     });
 
     it("should return the input when it is a valid url", () => {
       const url = "http://test.com";
-      expect(sanitizeUrl("Url", ["http:", "https:"])(url)).toStrictEqual(url);
+      expect(validateUrl("Url", url, ["http:", "https:"])).toStrictEqual(url);
     });
   });
 
-  describe("sanitizeHexString", () => {
+  describe("validateHexString", () => {
     it("should throw an error when the input is not a hex string", () => {
       const invalidHexString = "0a1f";
       const expectedLength = 2;
-      expect(() => sanitizeHexString("HexString", expectedLength)(invalidHexString)).toThrow(
+      expect(() => validateHexString("HexString", invalidHexString, expectedLength)).toThrow(
         `HexString must be hexadecimal string of length ${expectedLength}.`,
       );
     });
@@ -58,7 +60,7 @@ describe("CLI", () => {
     it("should throw an error when the input length is not equal to the expected length", () => {
       const hexString = "0x0a1f";
       const expectedLength = 4;
-      expect(() => sanitizeHexString("HexString", expectedLength)(hexString)).toThrow(
+      expect(() => validateHexString("HexString", hexString, expectedLength)).toThrow(
         `HexString must be hexadecimal string of length ${expectedLength}.`,
       );
     });
@@ -66,19 +68,7 @@ describe("CLI", () => {
     it("should return the input when it is a hex string", () => {
       const hexString = "0x0a1f";
       const expectedLength = 2;
-      expect(sanitizeHexString("HexString", expectedLength)(hexString)).toStrictEqual(hexString);
-    });
-  });
-
-  describe("sanitizeETHThreshold", () => {
-    it("should throw an error when the input threshold is less than 1 ETH", () => {
-      const invalidThreshold = "0.5";
-      expect(() => sanitizeETHThreshold()(invalidThreshold)).toThrow("Threshold must be higher than 1 ETH");
-    });
-
-    it("should return the input when it is higher than 1 ETH", () => {
-      const threshold = "2";
-      expect(sanitizeETHThreshold()(threshold)).toStrictEqual(threshold);
+      expect(validateHexString("HexString", hexString, expectedLength)).toStrictEqual(hexString);
     });
   });
 });
