+++ operations/src/utils/eth-transfer/constants.ts
@@ -1,0 +1,3 @@
+export const WEB3_SIGNER_PUBLIC_KEY_LENGTH = 64;
+export const DEFAULT_MAX_FEE_PER_GAS = "100000000000";
+export const DEFAULT_GAS_ESTIMATION_PERCENTILE = 10;
