+++ operations/tsconfig.json
@@ -1,16 +1,16 @@
 {
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "module": "NodeNext",
-        "moduleResolution": "NodeNext",
-        "target": "ES2022",
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "declaration": true,
-        "rootDir": "src",
-        "outDir": "dist",
-    },
-    "include": [
-        "src/**/*"
-    ],
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "module": "ES2022",
+    "outDir": "dist",
+    "rootDir": "src",
+    "strict": true,
+    "target": "es2022",
+    "moduleResolution": "Node",
+    "skipLibCheck": true,
+    "esModuleInterop": true,
+  },
+  "include": ["./src/**/*"],
+  "exclude": ["node_modules", "dist"]
 }
